Fix sprite loading so sprites are 32 pixels high
SpriteGrid was treating the sprite sheet as 6 sprites, each 32 pixels
wide and 64 pixels high. This was due to creating the SpriteGrid with
'rows' and 'sprite_count' set to None, so it defaulted to having a
single row.

The fix is to explictly state the sprite sheet has 2 rows of sprites.